Port the WIP 1.12.2 code to 1.19, part 9: Stuff that needs to go to 1.12.2

+++ src/main/java/fi/dy/masa/malilib/render/ShapeRenderUtils.java
@@ -207,7 +207,7 @@ public class ShapeRenderUtils
                                                   int width, int height,
                                                   BufferBuilder buffer)
     {
-        double pixelSize = 0.00390625; // 1 / 256
+        float pixelSize = 0.00390625F; // 1 / 256
         renderTexturedRectangle(x, y, z, u, v, width, height, pixelSize, pixelSize, buffer);
     }
 
@@ -219,7 +219,7 @@ public class ShapeRenderUtils
     public static void renderTexturedRectangle(double x, double y, double z,
                                                int u, int v,
                                                int renderWidth, int renderHeight,
-                                               double pixelWidth, double pixelHeight)
+                                               float pixelWidth, float pixelHeight)
     {
         renderScaledTexturedRectangle(x, y, z, u, v,
                                       renderWidth, renderHeight,
@@ -236,7 +236,7 @@ public class ShapeRenderUtils
     public static void renderTexturedRectangle(double x, double y, double z,
                                                int u, int v,
                                                int renderWidth, int renderHeight,
-                                               double pixelWidth, double pixelHeight,
+                                               float pixelWidth, float pixelHeight,
                                                BufferBuilder buffer)
     {
         renderScaledTexturedRectangle(x, y, z, u, v,
@@ -256,7 +256,7 @@ public class ShapeRenderUtils
                                                      int u, int v,
                                                      int renderWidth, int renderHeight,
                                                      int textureWidth, int textureHeight,
-                                                     double pixelWidth, double pixelHeight)
+                                                     float pixelWidth, float pixelHeight)
     {
         BufferBuilder buffer = RenderUtils.startBuffer(VertexFormat.DrawMode.QUADS, VertexFormats.POSITION_TEXTURE, true);
 
@@ -281,14 +281,14 @@ public class ShapeRenderUtils
                                                      int u, int v,
                                                      int renderWidth, int renderHeight,
                                                      int textureWidth, int textureHeight,
-                                                     double pixelWidth, double pixelHeight, BufferBuilder buffer)
+                                                     float pixelWidth, float pixelHeight, BufferBuilder buffer)
     {
         double x2 = x + renderWidth;
         double y2 = y + renderHeight;
-        double u1 = u                  * pixelWidth;
-        double u2 = (u + textureWidth) * pixelWidth;
-        double v1 = v                   * pixelHeight;
-        double v2 = (v + textureHeight) * pixelHeight;
+        float u1 = u                  * pixelWidth;
+        float u2 = (u + textureWidth) * pixelWidth;
+        float v1 = v                   * pixelHeight;
+        float v2 = (v + textureHeight) * pixelHeight;
 
         buffer.vertex(x , y2, z).texture(u1, v2).next();
         buffer.vertex(x2, y2, z).texture(u2, v2).next();
